Fix bugs in playlist logic code. Merge 'setTrackIds' and 'addOrRemoveTracks' methods into 'updatePlaylistTracks' that does it all.

+++ src/main/java/net/ehicks/loon/PlaylistLogic.java
@@ -10,6 +10,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
+import java.util.concurrent.atomic.AtomicLong;
 import java.util.stream.Collectors;
 
 @Service
@@ -26,11 +27,29 @@ public class PlaylistLogic
         this.trackRepo = trackRepo;
     }
 
-    // add or remove tracks from a playlist
-    public void addOrRemoveTracks(Playlist playlist, List<Long> trackIds, String action)
+    public void updatePlaylistTracks(Playlist playlist, List<Long> trackIds, String action, boolean clearExisting)
     {
+        if (clearExisting && trackIds.isEmpty())
+        {
+            playlistTrackRepo.deleteAll(playlistTrackRepo.findByPlaylistIdOrderByIndex(playlist.getId()));
+            playlist.setPlaylistTracks(new HashSet<>());
+            return;
+        }
+
+        if (clearExisting && action.equals("add") /* is 'action==add' necessary? */)
+        {
+            List<PlaylistTrack> playlistTracksNotInInput = playlist.getPlaylistTracks()
+                    .stream()
+                    .filter(playlistTrack -> trackIds.contains(playlistTrack.getTrack().getId()))
+                    .collect(Collectors.toList());
+
+            playlist.getPlaylistTracks().removeAll(playlistTracksNotInInput);
+            playlistTrackRepo.deleteAll(playlistTracksNotInInput);
+        }
+
         List<PlaylistTrack> tracksToPersist = new ArrayList<>();
 
+        AtomicLong nextId = new AtomicLong(getNextAvailableIndex(playlist.getId()));
         trackIds.forEach(trackId -> {
             PlaylistTrack playlistTrack = playlistTrackRepo.findByPlaylistIdAndTrackId(playlist.getId(), trackId);
             if (action.equals("add") && playlistTrack == null)
@@ -45,7 +64,7 @@ public class PlaylistLogic
                 playlistTrack = new PlaylistTrack();
                 playlistTrack.setPlaylist(playlist);
                 playlistTrack.setTrack(track);
-                playlistTrack.setIndex(getNextAvailableIndex(playlist.getId()));
+                playlistTrack.setIndex(nextId.getAndIncrement());
                 tracksToPersist.add(playlistTrack);
 
                 playlist.getPlaylistTracks().add(playlistTrack);
@@ -57,65 +76,8 @@ public class PlaylistLogic
             }
         });
 
-        if (action.equals("add"))
-            playlistTrackRepo.saveAll(tracksToPersist);
-        if (action.equals("remove"))
-            playlistTrackRepo.deleteAll(tracksToPersist);
-
-        consolidateTrackIndexes(playlist.getPlaylistTracks());
-    }
-
-    // replaces current tracks with newTrackIds
-    public void setTrackIds(Playlist playlist, List<Long> newTrackIds)
-    {
-        List<Long> currentTrackIds = playlist.getPlaylistTracks().stream().map(playlistTrack -> playlistTrack.getTrack().getId()).collect(Collectors.toList());
-        Set<PlaylistTrack> currentPlaylistTracks = playlist.getPlaylistTracks();
-
-        Set<PlaylistTrack> playlistTracksToRemove = new HashSet<>();
-        currentPlaylistTracks.forEach(currentPlaylistTrack -> {
-            if (!newTrackIds.contains(currentPlaylistTrack.getTrack().getId()))
-                playlistTracksToRemove.add(currentPlaylistTrack);
-        });
-
-        List<Long> trackIdsToRemove = new ArrayList<>(currentTrackIds);
-        trackIdsToRemove.removeAll(newTrackIds);
-
-        List<Long> trackIdsToAdd = new ArrayList<>(newTrackIds);
-        trackIdsToAdd.removeAll(currentTrackIds);
-
-        if (trackIdsToRemove.size() > 0)
-        {
-            playlist.getPlaylistTracks().removeAll(playlistTracksToRemove);
-            playlistTrackRepo.deleteAll(playlistTracksToRemove);
-        }
-
-        if (trackIdsToAdd.size() > 0)
-        {
-            List<Track> tracks = trackRepo.findAllByIdIn(trackIdsToAdd);
-
-            Set<PlaylistTrack> playlistTracksToAdd = new LinkedHashSet<>();
-            trackIdsToAdd.forEach(trackIdToAdd -> {
-                Track trackToAdd = tracks.stream().filter(track -> track.getId().equals(trackIdToAdd)).findFirst().orElse(null);
-                if (trackToAdd == null)
-                {
-                    log.error("Can't add track:" + trackIdToAdd + " to " + playlist);
-                    return;
-                }
-
-                PlaylistTrack playlistTrack = new PlaylistTrack();
-                playlistTrack.setPlaylist(playlist);
-                playlistTrack.setTrack(trackToAdd);
-                playlistTrack.setIndex(getNextAvailableIndex(playlist.getId()));
-                playlistTracksToAdd.add(playlistTrack);
-            });
-
-            if (playlistTracksToAdd.size() > 0)
-            {
-                playlistTrackRepo.saveAll(playlistTracksToAdd);
-
-                playlist.getPlaylistTracks().addAll(playlistTracksToAdd);
-            }
-        }
+        if (action.equals("add")) playlistTrackRepo.saveAll(tracksToPersist);
+        if (action.equals("remove")) playlistTrackRepo.deleteAll(tracksToPersist);
 
         consolidateTrackIndexes(playlist.getPlaylistTracks());
     }
@@ -125,7 +87,12 @@ public class PlaylistLogic
     {
         long nextIndex = 0;
 
-        for (PlaylistTrack playlistTrack : playlistTracks)
+        List<PlaylistTrack> tracks = new ArrayList<>(playlistTracks)
+                .stream()
+                .sorted(Comparator.comparing(PlaylistTrack::getIndex))
+                .collect(Collectors.toList());
+
+        for (PlaylistTrack playlistTrack : tracks)
         {
             if (!playlistTrack.getIndex().equals(nextIndex))
                 playlistTrack.setIndex(nextIndex);
